+++ Proline-DAL/src/main/scala/fr/proline/core/om/provider/lcms/IMapSetProvider.scala
@@ -4,6 +4,6 @@ import fr.proline.core.om.model.lcms.MapSet
 
 trait IMapSetProvider {
   
-  def getMapSet( mapSetId: Int ): MapSet
+  def getMapSet( mapSetId: Long ): MapSet
 
 }
