+++ Proline-DAL/src/main/scala/fr/proline/core/om/provider/msi/IPeptideInstanceProvider.scala
@@ -12,14 +12,14 @@ trait IPeptideInstanceProvider {
    *  @param pepInstIds: Sequence of ids of PeptideInstance to search for
    *  @return Array of Option[PeptideInstance] corresponding to found PeptideInstance
    */
-  def getPeptideInstancesAsOptions( pepInstIds: Seq[Int] ): Array[Option[PeptideInstance]]
+  def getPeptideInstancesAsOptions( pepInstIds: Seq[Long] ): Array[Option[PeptideInstance]]
   
   /**
    *  Get PeptideInstances with specified Ids.
    *  @param pepInstIds: Sequence of ids of PeptideInstance to search for
    *  @return Array of PeptideInstance corresponding to found PeptideInstance
    */
-  def getPeptideInstances( pepInstIds: Seq[Int] ): Array[PeptideInstance]
+  def getPeptideInstances( pepInstIds: Seq[Long] ): Array[PeptideInstance]
   
   /**
    *  Get PeptideInstance (wrapped in Option) with specified Id.
@@ -27,7 +27,7 @@ trait IPeptideInstanceProvider {
    *  @param pepInstId: id of PeptideInstance to search for
    *  @return Option[PeptideInstance] corresponding to found PeptideInstance
    */
-  def getPeptideInstance( pepInstId:Int ): Option[PeptideInstance] = {
+  def getPeptideInstance( pepInstId:Long ): Option[PeptideInstance] = {
     getPeptideInstancesAsOptions( Array(pepInstId) )(0)
   }
   
@@ -36,14 +36,14 @@ trait IPeptideInstanceProvider {
    *  @param resultSummaryIds: Sequence of ResultSummary ids to get PeptideInstance for
    *  @return Array of Option[PeptideInstance]belonging to specified one of the specified ResultSummary
    */
-  def getResultSummariesPeptideInstances( resultSummaryIds: Seq[Int] ): Array[PeptideInstance]
+  def getResultSummariesPeptideInstances( resultSummaryIds: Seq[Long] ): Array[PeptideInstance]
   
   /**
    *  Get PeptideInstance (wrapped in Option) associated to ResultSummary with specified Id.
    *  @param resultSummaryId: ResultSummary id to ger PeptideInstance for
    *  @return Array of Option[PeptideInstance]belonging to specified ResultSummary
    */
-  def getResultSummaryPeptideInstances( resultSummaryId: Int ): Array[PeptideInstance] = {
+  def getResultSummaryPeptideInstances( resultSummaryId: Long ): Array[PeptideInstance] = {
     getResultSummariesPeptideInstances( Array(resultSummaryId) )
   }
 }
