+++ Proline-DAL/src/main/scala/fr/proline/core/om/provider/msi/IPeptideMatchProvider.scala
@@ -5,24 +5,24 @@ import fr.proline.context.DatabaseConnectionContext
 
 trait IPeptideMatchProvider {
  
-  def getPeptideMatchesAsOptions( pepMatchIds: Seq[Int] ): Array[Option[PeptideMatch]]
+  def getPeptideMatchesAsOptions( pepMatchIds: Seq[Long] ): Array[Option[PeptideMatch]]
   
-  def getPeptideMatches( pepMatchIds: Seq[Int] ): Array[PeptideMatch]
+  def getPeptideMatches( pepMatchIds: Seq[Long] ): Array[PeptideMatch]
   
-  def getResultSetsPeptideMatches( resultSetIds: Seq[Int] ): Array[PeptideMatch]
+  def getResultSetsPeptideMatches( resultSetIds: Seq[Long] ): Array[PeptideMatch]
   
-  def getResultSummariesPeptideMatches( rsmIds: Seq[Int] ): Array[PeptideMatch]
+  def getResultSummariesPeptideMatches( rsmIds: Seq[Long] ): Array[PeptideMatch]
   
   
-  def getPeptideMatch( pepMatchId:Int ): Option[PeptideMatch] = {
+  def getPeptideMatch( pepMatchId:Long ): Option[PeptideMatch] = {
     getPeptideMatchesAsOptions( Array(pepMatchId) )(0)
   }
   
-  def getResultSetPeptideMatches( resultSetId: Int ): Array[PeptideMatch] = {
+  def getResultSetPeptideMatches( resultSetId: Long ): Array[PeptideMatch] = {
     getResultSetsPeptideMatches( Array(resultSetId) )
   }
   
-  def getResultSummaryPeptideMatches( rsmId: Int ): Array[PeptideMatch] = {
+  def getResultSummaryPeptideMatches( rsmId: Long ): Array[PeptideMatch] = {
     getResultSummariesPeptideMatches( Array(rsmId) )
   }
