+++ Proline-DAL/src/main/scala/fr/proline/core/om/provider/msi/IProteinMatchProvider.scala
@@ -5,24 +5,24 @@ import fr.proline.context.DatabaseConnectionContext
 
 trait IProteinMatchProvider {
   
-  def getProteinMatchesAsOptions( protMatchIds: Seq[Int] ): Array[Option[ProteinMatch]]
+  def getProteinMatchesAsOptions( protMatchIds: Seq[Long] ): Array[Option[ProteinMatch]]
 
-  def getProteinMatches( protMatchIds: Seq[Int] ): Array[ProteinMatch]
+  def getProteinMatches( protMatchIds: Seq[Long] ): Array[ProteinMatch]
   
-  def getResultSetsProteinMatches( resultSetIds: Seq[Int] ): Array[ProteinMatch]
+  def getResultSetsProteinMatches( resultSetIds: Seq[Long] ): Array[ProteinMatch]
   
-  def getResultSummariesProteinMatches( rsmIds: Seq[Int] ): Array[ProteinMatch]
+  def getResultSummariesProteinMatches( rsmIds: Seq[Long] ): Array[ProteinMatch]
   
   
-  def getProteinMatch( protMatchId:Int ): Option[ProteinMatch] = {
+  def getProteinMatch( protMatchId:Long ): Option[ProteinMatch] = {
     getProteinMatchesAsOptions( Array(protMatchId) )(0)
   }
   
-  def getResultSetProteinMatches( resultSetId: Int ): Array[ProteinMatch] = {
+  def getResultSetProteinMatches( resultSetId: Long ): Array[ProteinMatch] = {
     getResultSetsProteinMatches( Array(resultSetId) )
   }
   
-  def getResultSummaryProteinMatches( rsmId: Int ): Array[ProteinMatch] = {
+  def getResultSummaryProteinMatches( rsmId: Long ): Array[ProteinMatch] = {
     getResultSummariesProteinMatches( Array(rsmId) )
   }
