+++ Proline-DAL/src/main/scala/fr/proline/core/om/provider/msi/IProteinProvider.scala
@@ -14,7 +14,7 @@ trait IProteinProvider {
    *  @param protIds: Sequence of ids of Protein to search for
    *  @return Array of Option[Protein] corresponding to found Protein
    */
-  def getProteinsAsOptions( protIds: Seq[Int] ): Array[Option[Protein]]
+  def getProteinsAsOptions( protIds: Seq[Long] ): Array[Option[Protein]]
   
   /**
    * Get Protein (wrapped in Option) with specified Id.
@@ -23,7 +23,7 @@ trait IProteinProvider {
    *  @param protId: id of Protein to search for
    *  @return Option[Protein] corresponding to found Protein
    */
-  def getProtein( protId:Int ): Option[Protein] = getProteinsAsOptions( Array(protId) )(0) 
+  def getProtein( protId:Long ): Option[Protein] = getProteinsAsOptions( Array(protId) )(0) 
   
   /**
    * Get Protein (wrapped in Option) with specified sequence.
