+++ Proline-DAL/src/main/scala/fr/proline/core/om/provider/msi/IProteinSetProvider.scala
@@ -5,17 +5,17 @@ import fr.proline.context.DatabaseConnectionContext
 
 trait IProteinSetProvider {
   
-  def getProteinSetsAsOptions( protSetIds: Seq[Int] ): Array[Option[ProteinSet]]
+  def getProteinSetsAsOptions( protSetIds: Seq[Long] ): Array[Option[ProteinSet]]
   
-  def getProteinSets( protSetIds: Seq[Int] ): Array[ProteinSet]
+  def getProteinSets( protSetIds: Seq[Long] ): Array[ProteinSet]
   
-  def getProteinSet( protSetId:Int ): Option[ProteinSet] = {
+  def getProteinSet( protSetId:Long ): Option[ProteinSet] = {
     getProteinSetsAsOptions( Array(protSetId) )(0)
   }
   
-  def getResultSummariesProteinSets( resultSummaryIds: Seq[Int] ): Array[ProteinSet]
+  def getResultSummariesProteinSets( resultSummaryIds: Seq[Long] ): Array[ProteinSet]
   
-  def getResultSummaryProteinSets( resultSummaryId: Int ): Array[ProteinSet] = {
+  def getResultSummaryProteinSets( resultSummaryId: Long ): Array[ProteinSet] = {
     getResultSummariesProteinSets( Array(resultSummaryId) )
   }
 }
